requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.12.0
peft>=0.4.0
trl>=0.7.1
bitsandbytes>=0.39.0
accelerate>=0.20.3
wandb>=0.15.3
huggingface-hub>=0.15.1
python-dotenv>=1.0.0
pyyaml>=6.0
numpy>=1.24.3
pandas>=2.0.2
scikit-learn>=1.2.2
tqdm>=4.65.0
